refactor: split React components into separate files
- Extract PageTracker, Breadcrumbs, Header into layout components
- Move SkillsSection to dedicated sections folder
- Split HomePage, ResumeRequestPage, NotFoundPage into separate page files
- Simplify App.jsx from 442 lines to clean routing component
- Improve code maintainability and reusability
